+++ src/frontend/js/components/Search/index.spec.tsx
@@ -7,9 +7,9 @@ import { IntlProvider } from 'react-intl';
 
 import { History, HistoryContext } from 'data/useHistory';
 import * as mockWindow from 'utils/indirection/window';
+import { ContextFactory as mockContextFactory } from 'utils/test/factories';
 import createQueryClient from 'utils/react-query/createQueryClient';
-import { ContextFactory } from 'utils/test/factories';
-import { CommonDataProps } from 'types/commonDataProps';
+import context from 'utils/context';
 import Search from '.';
 
 let mockMatches = false;
@@ -22,6 +22,11 @@ jest.mock('utils/indirection/window', () => ({
     removeListener: jest.fn(),
   }),
   scroll: jest.fn(),
+}));
+
+jest.mock('utils/context', () => ({
+  __esModule: true,
+  default: mockContextFactory().generate(),
 }));
 
 describe('<Search />', () => {
@@ -37,7 +42,6 @@ describe('<Search />', () => {
     historyReplaceState,
   ];
   const queryClient = createQueryClient();
-  const contextProps: CommonDataProps['context'] = ContextFactory().generate();
 
   afterEach(() => {
     queryClient.clear();
@@ -57,7 +61,7 @@ describe('<Search />', () => {
       <QueryClientProvider client={queryClient}>
         <IntlProvider locale="en">
           <HistoryContext.Provider value={makeHistoryOf({ limit: '20', offset: '0' })}>
-            <Search context={contextProps} />
+            <Search context={context} />
           </HistoryContext.Provider>
         </IntlProvider>
       </QueryClientProvider>,
@@ -80,7 +84,7 @@ describe('<Search />', () => {
       <QueryClientProvider client={queryClient}>
         <IntlProvider locale="en">
           <HistoryContext.Provider value={makeHistoryOf({ limit: '20', offset: '0' })}>
-            <Search context={contextProps} />
+            <Search context={context} />
           </HistoryContext.Provider>
         </IntlProvider>
       </QueryClientProvider>,
@@ -110,7 +114,7 @@ describe('<Search />', () => {
       <QueryClientProvider client={queryClient}>
         <IntlProvider locale="en">
           <HistoryContext.Provider value={makeHistoryOf({ limit: '20', offset: '0' })}>
-            <Search context={contextProps} />
+            <Search context={context} />
           </HistoryContext.Provider>
         </IntlProvider>
       </QueryClientProvider>,
@@ -139,7 +143,7 @@ describe('<Search />', () => {
       <QueryClientProvider client={queryClient}>
         <IntlProvider locale="en">
           <HistoryContext.Provider value={makeHistoryOf({ limit: '20', offset: '0' })}>
-            <Search context={contextProps} />
+            <Search context={context} />
           </HistoryContext.Provider>
         </IntlProvider>
       </QueryClientProvider>,
@@ -207,7 +211,7 @@ describe('<Search />', () => {
                 lastDispatchActions: [{ type: 'FILTER_RESET' }],
               })}
             >
-              <Search context={contextProps} />
+              <Search context={context} />
             </HistoryContext.Provider>
           </IntlProvider>
         </QueryClientProvider>,
@@ -239,7 +243,7 @@ describe('<Search />', () => {
                 lastDispatchActions: [{ type: 'QUERY_UPDATE' }],
               })}
             >
-              <Search context={contextProps} />
+              <Search context={context} />
             </HistoryContext.Provider>
           </IntlProvider>
         </QueryClientProvider>,